leftbar, dropdown & navbar styling
