+++ frontend/src/App.tsx
@@ -10,12 +10,12 @@ import LeftBar from "./components/leftbar/LeftBar";
 import NavBar from "./components/navbar/NavBar";
 import Home from "./pages/home/Home";
 import Login from "./pages/login/Login";
-import UserProfile from "./pages/userProfile/UserProfile";
-import MyProfile from "./pages/myProfile/MyProfile";
+import UserProfile from "./pages/profile-pages/UserProfile";
+import MyProfile from "./pages/profile-pages/MyProfile";
 import Register from "./pages/register/Register";
 import Friends from "./pages/friends/Friends";
 import Chat from "./pages/chat/Chat";
-import ProfileUpdate from "./components/profileUpdate/ProfileUpdate";
+import ProfileUpdate from "./components/profile/profileUpdate/ProfileUpdate";
 import axios from "axios";
 import "./style.scss";
 import UserType from "./types/UserType";
