+++ frontend/src/components/profile/profileBanner/profileBanner.scss
