+++ frontend/src/components/profile/profileContent.scss
@@ -1,4 +1,4 @@
-@import "../../style.scss";
+
 
 .noPosts {
     display: flex;
