--- frontend/src/pages/myProfile/myProfile.scss
@@ -1,7 +1,0 @@
-.profilePage {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    padding-left: 1%;
-}
