+++ frontend/src/pages/profile-pages/MyProfile.tsx
@@ -1,9 +1,9 @@
 
-import "./myProfile.scss"
+import "./profilePage.scss"
 import ChangePassword from "../../components/changePassword/ChangePassword";
 import CreateTopic from "../../components/createTopic/CreateTopic";
 import DeleteAccount from "../../components/deleteAccount/DeleteAccount";
-import ProfileContent from "../../components/profileContent/ProfileContent";
+import ProfileContent from "../../components/profile/ProfileContent";
 import UserType from "../../types/UserType";
 //import CurrentUserProfile from "../../components/profile/CurrentUserProfile";
