+++ frontend/src/pages/profile-pages/profilePage.scss
